Implement interactive opengl renderer
